requirements.txt
aws-cdk-lib==2.8.0
constructs==10.0.34
boto3
